requirements.txt
python_pptx==1.0.2
selenium==4.34.2
tm1py==2.1
